+++ tests/ServicesBaseTest.php
@@ -9,6 +9,9 @@ require_once 'Modules/services/Controller/ServicesconfigController.php';
 require_once 'Modules/services/Controller/ServiceslistingController.php';
 require_once 'Modules/services/Controller/ServicesvisaController.php';
 require_once 'Modules/services/Controller/ServicesoriginsController.php';
+require_once 'Modules/services/Controller/ServicespurchaseController.php';
+require_once 'Modules/services/Controller/ServicesordersController.php';
+
 require_once 'Modules/services/Controller/ServicesprojectsController.php';
 require_once 'Modules/services/Controller/StockcabinetController.php';
 require_once 'Modules/services/Controller/StockshelfController.php';
@@ -76,6 +79,15 @@ class ServicesBaseTest extends BaseTest {
         $data = $c->editAction($space['id'], '');
         $this->assertTrue($data['service']['id'] > 0);
         return $data['service'];
+    }
+
+    protected function getServices($space) {
+        $req = new Request([
+            "path" => "services/".$space['id']
+        ], false);
+        $c = new ServiceslistingController($req, $space);
+        $data = $c->listingAction($space['id']);
+        return $data['services'];        
     }
 
     protected function createVisa($space, $user) {
@@ -167,6 +179,44 @@ class ServicesBaseTest extends BaseTest {
 
     }
 
+
+    protected function createPurchase($space, $service, $count) {
+        Configuration::getLogger()->debug('create puchase', ['service' => $service, 'space' => $space]);
+        $req = new Request([
+            "path" => "servicespurchaseedit/".$space['id']."/0",
+            "formid" => "editserviceform",
+            "comment" => "new purchase",
+            "date" => date('Y-m-d'),
+            "services" => [$service['id']],
+            "quantities" => [$count]
+
+        ], false);
+        $c = new ServicespurchaseController($req, $space);
+        $data = $c->editAction($space['id'], 0);
+        $this->assertTrue($data['purchase']['id'] > 0);
+        return $data['purchase'];
+    }
+
+    protected function createOrder($space, $service, $user, $client, $count) {
+        Configuration::getLogger()->debug('create puchase', ['service' => $service, 'space' => $space]);
+        $req = new Request([
+            "path" => "servicesorderedit/".$space['id']."/0",
+            "formid" => "orderEditForm",
+            "no_identification" => 'order'.time(),
+            "id_user" => $user['id'],
+            "id_client" => $client['id'],
+            "id_status" => 1,
+            "date_open" => date('Y-m-d'),
+            "date_close" => '',
+            "services" => [$service['id']],
+            "quantities" => [$count]
+
+        ], false);
+        $c = new ServicesordersController($req, $space);
+        $data = $c->editAction($space['id'], 0);
+        $this->assertTrue($data['order']['id'] > 0);
+        return $data['order'];
+    }
 
 }